debugged mcparticle tree. supera runs.  however, only neutrino interaction used for interaction image or semantic segmentation image
