+++ app/Supera/APILArLite/LinkDef.h
@@ -11,6 +11,7 @@
 
 #pragma link C++ namespace larcv::supera;
 //#pragma link C++ class larlite::Supera+;
+#pragma link C++ class larcv::supera::MCPNode+;
 #pragma link C++ class larcv::supera::MCParticleTree<larlite::mctruth,larlite::mctrack,larlite::mcshower,larlite::simch>+;
 #pragma link C++ class larcv::supera::Cropper<larlite::mctrack,larlite::mcshower,larlite::simch>+;
