+++ app/Supera/larfmwk_shared/MCPNode.cxx
@@ -1,0 +1,1 @@
+#include "MCPNode.h"
